+++ NEWS.md
@@ -1,3 +1,10 @@
+# NicheScribe 0.2.0
+
+* Updated NicheScribe reference.
+* Addressed future.globals.maxSize bug.
+* Improved handling of reference data.
+* Updated example in README.
+
 # NicheScribe 0.1.0
 
 * Initial code commit.